+++ CHANGELOG.md
@@ -7,6 +7,14 @@
 
 ### New features
 
+### Bugfixes
+
+</details>
+
+## 0.3.1
+
+### New features
+
  - Make `\WizaplaceFrontBundle\Twig\AppExtension::formatPrice` use the `currency` translation key instead of `€`.
  - Upgrade SDK to v1.15.1
  - Added `\WizaplaceFrontBundle\Service\JsonSearchService`
@@ -14,8 +22,6 @@
 ### Bugfixes
 
  - Fix a bug causing us to try to merge a basket with itself
-
-</details>
 
 ## 0.3.0